backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-redis==5.0.0
djangorestframework==3.12.4
future==0.18.2
hashids==1.3.1
idna==3.2
mysqlclient==2.0.3
pytz==2021.1
redis==3.5.3
serializers==0.2.4
sqlparse==0.4.1
urllib3==1.26.6
